pyproject.toml
[tool.poetry]
name = "invoice_manager"
version = "0.1.0"
description = ""
authors = ["hayaa1234 <hayalawansah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
reportlab = "^3.6.11"
requests = "^2.28.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
